requirements.txt
toml
mysql-connector-python
boto3
